Update progress reporting

+++ SessionTools/sort_sessions.py
@@ -67,14 +67,14 @@ skipped = 0
 for path in paths_to_sort_list:
     i += 1
     if i % 1000 == 0:
-        log ("Sorted: " + str(i) + ", " + str((100*i) / len(paths_to_sort)) + ", skipped: " + str(skipped))
+        log ("Sorted: " + str(i) + ", " + str(float((100*i)) / len(paths_to_sort)) + ", skipped: " + str(skipped))
     out_path = compute_derived_path(path)
     # skip files that have been processed already between start of the script and now
     if os.path.exists(out_path):
         skipped += 1
         continue
     try:
-        log(str((100*i) / len(paths_to_sort)) + "%: " + path)
+        log(str(float((100*i)) / len(paths_to_sort)) + "%: " + path)
         f = gzip.open(path)
         data_set = set()
         data = []
